+++ Yumemi-iOS-training/Parameter.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct Parameter: Encodable {
+    let area: String
+    let date: String
+}
