+++ Yumemi-iOS-training/Weather.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+struct Weather: Decodable {
+    let weather: String
+    let max_temp: Int
+    let min_temp: Int
+    let date: String
+}
